Updates to YGOCard component
